Отрисовала полноэкранное фото, открываются все фото, кроме первого
