+++ js/main.js
@@ -3,4 +3,5 @@ import './utils.js';
 import './pattern-photo.js';
 import './rendering-photo.js';
 import './fullsize-photo.js';
-import './fullsize-photo-actions.js';
+import './fullsize-photo-close.js';
+import './fullsize-photo-open.js';
